fix(fixed): remove business logic from Components to Hooks
