+++ src/screens/Bookmarks/hooks/useBookmarks.tsx
@@ -1,5 +1,5 @@
 import {useCallback} from 'react';
-import {useDispatch} from 'react-redux';
+import {useDispatch, useSelector} from 'react-redux';
 import {
   useRequestError,
   useRequestLoading,
@@ -9,11 +9,14 @@ import {getInitialHomeDataRequest} from 'core/reducers';
 import {IBookmarkItem} from 'core/types';
 import {useNavigation} from '@react-navigation/native';
 import {ObjectsListScreenNavigationProps} from '../types';
+import {selectBookmarksCardsData} from 'core/selectors';
 
 export const useBookmarks = () => {
   const navigation = useNavigation<ObjectsListScreenNavigationProps>();
   const dispatch = useDispatch();
   const {sendSelectSavedCategoryEvent} = useBookmarksAnalytics();
+
+  const bookmarksCategories = useSelector(selectBookmarksCardsData);
 
   const getHomeData = useCallback(() => {
     dispatch(getInitialHomeDataRequest());
@@ -34,6 +37,7 @@ export const useBookmarks = () => {
   );
 
   return {
+    bookmarksCategories,
     getHomeData,
     loading,
     error,
