+++ src/screens/CheckEmail/CheckEmail.tsx
@@ -1,50 +1,13 @@
-import React, {useCallback, useEffect} from 'react';
-import {useOnRequestSuccess, useTranslation} from 'core/hooks';
+import React from 'react';
+import {useTranslation} from 'core/hooks';
 import {FormInput} from 'atoms';
-import {checkUserEmailRequest} from 'core/reducers';
 import {AuthForm} from 'organisms';
 import {useCheckEmail} from './hooks';
 
 export const CheckEmail = () => {
   const {t} = useTranslation('authentification');
-  const {
-    navigation,
-    email,
-    setIsEmailCorrect,
-    dispatch,
-    isEmailCorrect,
-    loading,
-    setEmail,
-  } = useCheckEmail();
-
-  useEffect(() => {
-    const regexForEmail = /^[^\s@]+@[^\s@]+\.[^\s@]{2,}$/;
-
-    if (regexForEmail.test(email)) {
-      setIsEmailCorrect(true);
-    } else {
-      setIsEmailCorrect(false);
-    }
-  }, [email, setIsEmailCorrect]);
-
-  const checkEmail = useCallback(() => {
-    dispatch(checkUserEmailRequest(email));
-  }, [dispatch, email]);
-
-  useOnRequestSuccess(checkUserEmailRequest, data => {
-    if (!data.exist) {
-      navigation.navigate('SignUpForm', {email});
-    } else if (data.exist) {
-      if (data.isConfirmed) {
-        navigation.navigate('SignInPassword', {email});
-      } else {
-        navigation.navigate('EmailValidation', {
-          email,
-          isSignUp: data.isPasswordReset ? false : true,
-        });
-      }
-    }
-  });
+  const {email, isEmailCorrect, loading, setEmail, checkEmail} =
+    useCheckEmail();
 
   return (
     <AuthForm
