+++ src/screens/NewPassword/NewPassword.tsx
@@ -1,10 +1,5 @@
 import React from 'react';
-import {confirmNewPasswordRequest} from 'core/reducers';
-import {
-  useOnRequestSuccess,
-  useTogglePasswordVisibility,
-  useTranslation,
-} from 'core/hooks';
+import {useTranslation} from 'core/hooks';
 import {FormInput} from 'atoms';
 import {AuthForm} from 'organisms';
 import {useNewPassword} from './hooks';
@@ -12,20 +7,15 @@ import {useNewPassword} from './hooks';
 export const NewPassword = () => {
   const {t} = useTranslation('authentification');
   const {
-    navigation,
     onConfirmNewPassword,
     newPassword,
     loading,
     setNewPassword,
+    buttonText,
+    rightIcon,
+    passwordVisibility,
+    handlePasswordVisibility,
   } = useNewPassword();
-
-  const {passwordVisibility, rightIcon, handlePasswordVisibility} =
-    useTogglePasswordVisibility('eye');
-  const buttonText = t('save').toUpperCase();
-
-  useOnRequestSuccess(confirmNewPasswordRequest, () => {
-    navigation.getParent()?.goBack();
-  });
 
   return (
     <AuthForm
