+++ src/screens/ProfileSettingsTheme/hooks/useProfileSettingsTheme.tsx
@@ -1,10 +1,12 @@
 import {setTheme} from 'core/reducers';
-import {useDispatch} from 'react-redux';
+import {useDispatch, useSelector} from 'react-redux';
 import {THEME_TYPE} from 'core/constants';
 import {useCallback} from 'react';
+import {selectAppTheme} from 'core/selectors';
 
 export const useProfileSettingsTheme = () => {
   const dispatch = useDispatch();
+  const userTheme = useSelector(selectAppTheme);
 
   const changeTheme = useCallback(
     (theme: THEME_TYPE | null) => {
@@ -13,5 +15,5 @@ export const useProfileSettingsTheme = () => {
     [dispatch],
   );
 
-  return {changeTheme};
+  return {changeTheme, userTheme};
 };
