+++ src/screens/Splash/Splash.tsx
@@ -1,9 +1,15 @@
-import React, {useEffect} from 'react';
+import React, {useCallback, useEffect} from 'react';
 
 import {View} from 'react-native';
 import Animated, {interpolate, useAnimatedStyle} from 'react-native-reanimated';
 import {isIOS} from 'services/PlatformService';
-import {useSplash} from './hooks';
+import {
+  useSharedValue,
+  withTiming,
+  Easing,
+  runOnJS,
+} from 'react-native-reanimated';
+import RNBootSplash from 'react-native-bootsplash';
 import {styles} from './styles';
 
 interface IProps {
@@ -12,7 +18,64 @@ interface IProps {
 }
 
 export const Splash = ({onAnimationEnd, onFadeStart}: IProps) => {
-  const {opacity, animatedValue, animateIOS, animateAndroid} = useSplash();
+  const opacity = useSharedValue(1);
+  const animatedValue = useSharedValue(0);
+
+  const animateIOS = useCallback(
+    onAnimationEnd => {
+      setTimeout(() => {
+        RNBootSplash.hide().then(() => {
+          animatedValue.value = withTiming(
+            1,
+            {
+              duration: 300,
+              easing: Easing.out(Easing.ease),
+            },
+            () => {
+              if (onAnimationEnd) {
+                runOnJS(onAnimationEnd)();
+              }
+            },
+          );
+        });
+      }, 10);
+    },
+    [animatedValue],
+  );
+
+  const animateAndroid = useCallback(
+    (onFadeStart, onAnimationEnd) => {
+      setTimeout(() => {
+        RNBootSplash.hide().then(() => {
+          animatedValue.value = withTiming(
+            1,
+            {
+              duration: 400,
+              easing: Easing.out(Easing.ease),
+            },
+            () => {
+              if (onFadeStart) {
+                runOnJS(onFadeStart)();
+              }
+              opacity.value = withTiming(
+                0,
+                {
+                  duration: 300,
+                  easing: Easing.out(Easing.ease),
+                },
+                () => {
+                  if (onAnimationEnd) {
+                    runOnJS(onAnimationEnd)();
+                  }
+                },
+              );
+            },
+          );
+        });
+      }, 300);
+    },
+    [animatedValue, opacity],
+  );
 
   const containerAnimatedStyle = useAnimatedStyle(() => {
     return {
