--- src/screens/Splash/hooks/index.tsx
@@ -1,1 +1,0 @@
-export {useSplash} from './useSplash';
